Conflicts:
	qrgui/mainwindow/mainWindow.cpp
	qrgui/qrgui.pri

+++ qrgui/umllib/private/sceneGridHandler.h
@@ -76,13 +76,13 @@ private:
 	QPen mGuidesPen;
 
 	/** @brief radius of guides appearing */
-	static qreal const radius = 20;
+	static int const radius = 20;
 	/** @brief minimum distance to approach to jump to the guide */
-	static qreal const radiusJump = 10;
+	static int const radiusJump = 10;
 	/** @brief spacing between element and guide */
-	static qreal const spacing = 0;
+	static int const spacing = 0;
 	/** @brief guides that have distance equal or less will be represented by only one of them */
-	static qreal const indistinguishabilitySpace = 10;
+	static int const indistinguishabilitySpace = 10;
 };
 
 }
