+++ qrutils/mathUtils/gaussNoise.h
@@ -31,7 +31,7 @@ public:
 
 private:
 	static const int defaultApproximationLevel = 12;
-	static const qreal defaultDispersion = 1.0;
+	static const int defaultDispersion = 1;
 
 	/// Body of function 'generate'. Uses for various realizations of 'generate'
 	qreal genBody(unsigned int const approximationLevel, qreal const variance) const;
